Extract handleSocketConnection to a controller
